uncircularize dependency
