+++ src/configTypes.ts
@@ -1,0 +1,6 @@
+export interface GlobalConfig {
+  hasRun?: boolean;
+  account?: {
+    email?: string;
+  };
+}
